+++ Sources/DTModelStorage/BaseStorage.swift
@@ -31,15 +31,57 @@ public let DTTableViewElementSectionHeader = "DTTableViewElementSectionHeader"
 /// Suggested supplementary kind for UITableView footer
 public let DTTableViewElementSectionFooter = "DTTableViewElementSectionFooter"
 
-/// Base class for storage classes
-open class BaseStorage: NSObject
-{
+/// `BaseSupplementaryStorage` is a base class, that implements common functionality for `SupplementaryStorage` protocol and serves as a base class for `MemoryStorage`, `CoreDataStorage`, `SingleSectionStorage`, `RealmStorage`.
+open class BaseSupplementaryStorage: NSObject, SupplementaryStorage {
+    
+    /// Returns a header model for specified section index or nil.
+    open var headerModelProvider: ((Int) -> Any?)?
+    
+    /// Returns a footer model for specified section index or nil
+    open var footerModelProvider: ((Int) -> Any?)?
+
+    private var _supplementaryModelProvider: ((String, IndexPath) -> Any?)?
+    
+    /// Returns supplementary model for specified section indexPath and supplementary kind, or nil. Setter for this property is overridden to allow calling `headerModelProvider` and `footerModelProvider` closures.
+    open var supplementaryModelProvider: ((String, IndexPath) -> Any?)? {
+        get {
+            return _supplementaryModelProvider
+        }
+        set {
+            _supplementaryModelProvider = { [weak self] kind, indexPath in
+                if let headerModel = self?.headerModelProvider, self?.supplementaryHeaderKind == kind {
+                    return headerModel(indexPath.section)
+                }
+                if let footerModel = self?.footerModelProvider, self?.supplementaryFooterKind == kind {
+                    return footerModel(indexPath.section)
+                }
+                return newValue?(kind, indexPath)
+            }
+        }
+    }
+    
     /// Supplementary kind for header in current storage
     open var supplementaryHeaderKind: String?
     
     /// Supplementary kind for footer in current storage
     open var supplementaryFooterKind: String?
+}
+
+/// `StorageUpdating` protocol is used to transfer data storage updates.
+public protocol StorageUpdating : class
+{
+    /// Transfers data storage updates.
+    ///
+    /// Object, that implements this method, may react to received update by updating UI for current storage.
+    func storageDidPerformUpdate(_ update: StorageUpdate)
     
+    /// Method is called when UI needs to be fully updated for data storage changes.
+    func storageNeedsReloading()
+}
+
+/// Base class for storage classes
+open class BaseUpdateDeliveringStorage: BaseSupplementaryStorage
+{
     /// Current update
     open var currentUpdate: StorageUpdate?