+++ Sources/DTModelStorage/MemoryStorage.swift
@@ -139,7 +139,7 @@ public enum MemoryStorageError: LocalizedError
 /// `MemoryStorage` stores data models using array of `SectionModel` instances. It has various methods for changing storage contents - add, remove, insert, replace e.t.c.
 /// - Note: It also notifies it's delegate about underlying changes so that delegate can update interface accordingly
 /// - SeeAlso: `SectionModel`
-open class MemoryStorage: BaseStorage, Storage, SupplementaryStorage, SectionLocationIdentifyable, HeaderFooterSettable
+open class MemoryStorage: BaseUpdateDeliveringStorage, Storage, SectionLocationIdentifyable
 {
     /// When enabled, datasource updates are not applied immediately and saved inside `StorageUpdate` `enqueuedDatasourceUpdates` property.
     /// Call `StorageUpdate.applyDeferredDatasourceUpdates` method to apply all deferred changes.
@@ -202,61 +202,6 @@ open class MemoryStorage: BaseStorage, Storage, SupplementaryStorage, SectionLoc
         guard indexPath.section < sections.count else { return nil }
         guard indexPath.item < sections[indexPath.section].numberOfItems else { return nil }
         return sections[indexPath.section].item(at: indexPath.item)
-    }
-    
-    /// Sets section header `model` for section at `sectionIndex`
-    ///
-    /// This method calls delegate?.storageNeedsReloading() method at the end, causing UI to be updated.
-    /// - SeeAlso: `configureForTableViewUsage`
-    /// - SeeAlso: `configureForCollectionViewUsage`
-    open func setSectionHeaderModel<T>(_ model: T?, forSection sectionIndex: Int)
-    {
-        guard let headerKind = supplementaryHeaderKind else {
-            assertionFailure("supplementaryHeaderKind property was not set before calling setSectionHeaderModel: forSectionIndex: method"); return
-        }
-        let section = getValidSection(sectionIndex, collectChangesIn: nil)
-        section.setSupplementaryModel(model, forKind: headerKind, atIndex: 0)
-        delegate?.storageNeedsReloading()
-    }
-    
-    /// Sets section footer `model` for section at `sectionIndex`
-    ///
-    /// This method calls delegate?.storageNeedsReloading() method at the end, causing UI to be updated.
-    /// - SeeAlso: `configureForTableViewUsage`
-    /// - SeeAlso: `configureForCollectionViewUsage`
-    open func setSectionFooterModel<T>(_ model: T?, forSection sectionIndex: Int)
-    {
-        guard let footerKind = supplementaryFooterKind else {
-            assertionFailure("supplementaryFooterKind property was not set before calling setSectionFooterModel: forSectionIndex: method"); return
-        }
-        let section = getValidSection(sectionIndex, collectChangesIn: nil)
-        section.setSupplementaryModel(model, forKind: footerKind, atIndex: 0)
-        delegate?.storageNeedsReloading()
-    }
-    
-    /// Sets supplementary `models` for supplementary of `kind`.
-    ///
-    /// - Note: This method can be used to clear all supplementaries of specific kind, just pass an empty array as models.
-    open func setSupplementaries(_ models: [[Int: Any]], forKind kind: String)
-    {
-        defer {
-            self.delegate?.storageNeedsReloading()
-        }
-        
-        if models.count == 0 {
-            for index in 0..<self.sections.count {
-                let section = self.sections[index] as? SupplementaryAccessible
-                section?.supplementaries[kind] = nil
-            }
-            return
-        }
-        
-        _ = getValidSection(models.count - 1, collectChangesIn: nil)
-        
-        for index in 0 ..< models.count {
-            let section = self.sections[index] as? SupplementaryAccessible
-            section?.supplementaries[kind] = models[index]
-        }
     }
     
     /// Sets `items` for section at `index`.
@@ -706,18 +651,5 @@ open class MemoryStorage: BaseStorage, Storage, SupplementaryStorage, SectionLoc
         let unsorted = NSMutableArray(array: indexPaths)
         let descriptor = NSSortDescriptor(key: "self", ascending: ascending)
         return unsorted.sortedArray(using: [descriptor]) as? [IndexPath] ?? []
-    }
-    
-    // MARK: - SupplementaryStorage
-    
-    /// Returns supplementary model of supplementary `kind` for section at `sectionIndexPath`. Returns nil if not found.
-    ///
-    /// - SeeAlso: `headerModelForSectionIndex`
-    /// - SeeAlso: `footerModelForSectionIndex`
-    open func supplementaryModel(ofKind kind: String, forSectionAt sectionIndexPath: IndexPath) -> Any? {
-        guard sectionIndexPath.section < sections.count else {
-            return nil
-        }
-        return (self.sections[sectionIndexPath.section] as? SupplementaryAccessible)?.supplementaryModel(ofKind:kind, atIndex: sectionIndexPath.item)
     }
 }